+++ app/(afterLogin)/layout.tsx
@@ -1,5 +1,4 @@
 import { ReactNode } from 'react';
-import dynamic from 'next/dynamic';
 
 import RedirectToLogin from '@/app/(afterLogin)/_components/RedirectToLogin';
 import NavigationBar from '@/app/(afterLogin)/_components/NavigationBar';
@@ -7,11 +6,6 @@ import RQProvider from '@/app/(afterLogin)/_components/RQProvider';
 import Header from '@/app/(afterLogin)/_components/Header';
 
 type Props = { children: ReactNode };
-
-const FollowModal = dynamic(
-  () => import('@/app/(afterLogin)/users/[userId]/_components/FollowModal'),
-  { ssr: false },
-);
 
 const AfterLoginLayout = ({ children }: Props) => {
   return (
@@ -22,7 +16,6 @@ const AfterLoginLayout = ({ children }: Props) => {
         {children}
       </div>
       <NavigationBar />
-      <FollowModal />
     </RQProvider>
   );
 };
